requirements.txt
django==3.0.7
django-tastypie
oss2
six